# src/main/resources/static/swagger/index.json
{
	"swagger": "2.0",
	"info": {
		"description": "openapi",
		"version": "3.0.0",
		"title": "益友会",
		"termsOfService": "https://www.yestae.com/home"
	},
	"host": "openapi-test.yestae.com",
	"basePath": "/api",
...
	"paths": {
...
		"/sms/generateCode": {
			"post": {
				"tags": ["sms-api-controller"],
				"summary": "生成微信登录验证码",
				"description": "生成微信登录验证码",
				"operationId": "generateCodeUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"mobile\":\"18311112223\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/sms/sendAuthCode": {
			"post": {
				"tags": ["sms-api-controller"],
				"summary": "发送登录验证码",
				"description": "发送登录验证码",
				"operationId": "sendLoginCodeUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"mobile\":\"18311112223\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/sms/sendForgetPwdCode": {
			"post": {
				"tags": ["sms-api-controller"],
				"summary": "发送忘记密码短信验证码",
				"description": "发送忘记密码短信验证码",
				"operationId": "sendUpdatePasswordCodeUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"mobile\":\"18311112223\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/sms/sendUpdateMobileCode": {
			"post": {
				"tags": ["sms-api-controller"],
				"summary": "发送修改手机号验证码",
				"description": "发送修改手机号验证码",
				"operationId": "sendUpdateMobileCodeUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"mobile\":\"18311112223\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/sms/validatorSmsCode": {
			"post": {
				"tags": ["sms-api-controller"],
				"summary": "校验短信验证码",
				"description": "校验短信验证码",
				"operationId": "validatorSmsCodeUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"mobile\":\"18311112223\",\"smsCode\":\"1234\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
...
	},
	"securityDefinitions": {
		"token": {
			"type": "apiKey",
			"name": "token",
			"in": "header"
		}
	}
}